fix(KEN-815): a switch counts as code, and a code span is a mention
Six review rounds all found one class. Every reading of what a line hands
a program had a shape it did not cover, and each fix added more reading
for the next round to find a hole in. So the reading goes.

What is left answers the issue and closes the class by construction. A
markdown code span is a document naming a switch, and that is the one
quotation the rule reads. Everywhere else the switch counts, in any
language, which is what main did. No shape of a comment, a case arm, an
eval or a launcher can reopen the question, because none of them is asked
any more.

Deleted: the tokenizer promoted out of the fetch rule, which is back where
it was and serves that rule alone again; the quoted-argument and command
string readings; the case-arm cut and the shell comment scan; the shell,
opaque and prose three-way classification, which now has nothing left to
decide but whether a file is markdown.

Kept: the prose reading. Code spans read for the whole document, so a run
of backticks closing on a later line quotes what stands between; a run
reaching no match quotes nothing; a run stops at the end of its block, and
a heading, a setext underline, a list marker and a blockquote each end
one. An escaped backtick is a character. The line an indented block opens
on is the block's text. A .disabled suffix comes off before the file is
judged.

The cost, measured over the 19 skills this repository ships. growth-guards
goes from 68 to 71: its README and SKILL.md name the switch in backticks
three times and those are clean, while the two messages its hooks print
still rate Critical. orch stays at 78, unchanged from main, because the
eight lines of its tests that spell the permission switch inside a shell
string count as code. Both numbers are pinned to the real trees in
crates/core/tests/quality/corpus.rs.
