+++ changelog.d/fixed/KEN-815.md
@@ -1,3 +1,3 @@
-- The safety score counts a dangerous switch where a line would run it: a
-  README or `case` arm naming one no longer scores, and one on a command
-  line, quoted or in a Python or Rust file, still does.
+- The safety score no longer flags a dangerous switch a document writes in
+  a code span, so a README or SKILL.md explaining one scores clean. The
+  switch written as code still counts.
