+++ crates/core/src/quality/rules/content/fetch/tokens.rs
@@ -1,12 +1,109 @@
-//! Reading one line into the commands it holds. Quoting, escapes and
-//! substitutions are all answered here, in one pass over the characters,
-//! because every one of them changes where the next word or command
-//! begins.
+//! A command line as the shell would hand it over.
+//!
+//! Split out of `fetch.rs`, which is the rule that needs it. Which command
+//! runs the payload, which of its arguments is the download, where one
+//! command ends and the next begins — every one of those is a question
+//! about shell syntax, and every one of them has been answered here with a
+//! better search of the raw text and then defeated by the next piece of
+//! syntax: case, an address belonging to another command, an option that
+//! takes a value, a second fetch, a separator inside quotes. Reading the
+//! line once is one answer to all of them.
+//!
+//! Syntax only. Nothing here expands a variable, runs a substitution or
+//! matches a glob, so `curl $URL | sh` names `$URL` — which is what the
+//! line says, and what a person reading the finding will see in the file.
 
-use super::{Command, Reached, Word};
+/// One word, as the shell would pass it: quotes and escapes taken out.
+pub(super) struct Word {
+    pub(super) text: String,
+    /// Whether any of it was written inside quotes. A quoted word is an
+    /// operand the shell hands over exactly as it stands; a bare one is a
+    /// word in prose, and may have picked up punctuation from the sentence
+    /// around it.
+    pub(super) quoted: bool,
+}
+
+/// How a command was reached from the one before it.
+#[derive(PartialEq, Eq)]
+pub(super) enum Reached {
+    /// First on the line.
+    First,
+    /// `|` — the command before it writes into this one.
+    Pipe,
+    /// `&&` — this one runs if the command before it succeeded.
+    And,
+    /// `;`, `&`, `||` — this one runs whatever happened before it.
+    Next,
+}
+
+/// One command on a line: how it was reached, the words it is given, and
+/// where it sits in the line it came from.
+pub(super) struct Command {
+    pub(super) reached_by: Reached,
+    pub(super) at: std::ops::Range<usize>,
+    pub(super) words: Vec<Word>,
+}
+
+impl Command {
+    /// Where the program name sits: the first word that is not a variable
+    /// assignment. `MODE=x sh` runs `sh` with `MODE` set for it, and
+    /// reading the assignment as the program name misses the command
+    /// entirely — which for a runner means going quiet on a line that
+    /// pipes a download into a shell.
+    fn names_program(&self) -> Option<usize> {
+        self.words.iter().position(|word| !assigns(word))
+    }
+
+    /// The program being run, lowercased and by its own name rather than
+    /// the path it was reached through. `None` for a command that is
+    /// nothing but assignments, or an empty stretch between separators.
+    ///
+    /// `/bin/sh` and `./bash` run the same programs `sh` and `bash` do, and
+    /// a line piping a download into one of them is the thing this rule
+    /// exists to catch — matching the whole word means the rule says
+    /// nothing about it. The cut is at the last separator, so `notbash` and
+    /// `mybash.txt` are still their own whole names and still match
+    /// nothing. An address is not a path to a program, whatever it ends in.
+    pub(super) fn verb(&self) -> Option<String> {
+        let written = &self.words.get(self.names_program()?)?.text;
+        let named = match written.contains("://") {
+            true => written.as_str(),
+            false => written.rsplit('/').next().unwrap_or(written),
+        };
+        Some(named.to_ascii_lowercase())
+    }
+
+    pub(super) fn has_word(&self, word: &str) -> bool {
+        self.words.iter().any(|held| held.text == word)
+    }
+
+    /// Everything after the program name, as the shell would pass it —
+    /// counted from the program and not from the start of the command, or
+    /// an assignment before it shifts every argument by one.
+    pub(super) fn arguments(&self) -> &[Word] {
+        let Some(at) = self.names_program() else {
+            return &[];
+        };
+        self.words.get(at + 1..).unwrap_or_default()
+    }
+}
+
+/// Whether this word sets a variable for the command rather than being one:
+/// a name the shell would accept, then `=`. `--referer=https://x` is not
+/// one — the name has to be a name.
+fn assigns(word: &Word) -> bool {
+    let Some((name, _)) = word.text.split_once('=') else {
+        return false;
+    };
+    let mut letters = name.chars();
+    letters
+        .next()
+        .is_some_and(|first| first.is_ascii_alphabetic() || first == '_')
+        && letters.all(|c| c.is_ascii_alphanumeric() || c == '_')
+}
 
 /// The commands one line holds, in the order they are written.
-pub(in crate::quality) fn commands(line: &str) -> Vec<Command> {
+pub(super) fn commands(line: &str) -> Vec<Command> {
     let mut found = Vec::new();
     let mut state = Scan::default();
     let mut chars = line.char_indices().peekable();
@@ -109,10 +206,6 @@ struct Scan {
     /// Whether any of the word being read was written inside quotes.
     quoted: bool,
     start: Option<usize>,
-    /// Where the word being read begins, which is not where the command
-    /// does: a command keeps its first word's start and every word after
-    /// it needs one of its own.
-    word_start: Option<usize>,
     end: usize,
     words: Vec<Word>,
     reached_by: Option<Reached>,
@@ -128,7 +221,6 @@ impl Default for Scan {
             started: false,
             quoted: false,
             start: None,
-            word_start: None,
             end: 0,
             words: Vec::new(),
             reached_by: None,
@@ -231,7 +323,6 @@ impl Scan {
         if self.start.is_none() {
             self.start = Some(at);
         }
-        self.word_start.get_or_insert(at);
         self.started = true;
         self.quoted = true;
         self.end = at + c.len_utf8();
@@ -241,7 +332,6 @@ impl Scan {
         if self.start.is_none() {
             self.start = Some(at);
         }
-        self.word_start.get_or_insert(at);
         self.started = true;
         self.quoted |= self.inside();
         self.word.push(c);
@@ -258,7 +348,6 @@ impl Scan {
     fn end_word(&mut self) {
         let word = std::mem::take(&mut self.word);
         let quoted = std::mem::take(&mut self.quoted);
-        let at = std::mem::take(&mut self.word_start).unwrap_or(self.end)..self.end;
         if !std::mem::take(&mut self.started) {
             return;
         }
@@ -277,7 +366,7 @@ impl Scan {
         if !quoted && text.is_empty() {
             return;
         }
-        self.words.push(Word { text, quoted, at });
+        self.words.push(Word { text, quoted });
     }
 
     fn end_command(&mut self, reached: Reached, found: &mut Vec<Command>) {
@@ -302,3 +391,7 @@ impl Scan {
         });
     }
 }
+
+#[cfg(test)]
+#[allow(clippy::unwrap_used, clippy::expect_used)]
+mod tests;
