+++ crates/core/tests/quality/main.rs
@@ -12,7 +12,6 @@ mod fixture;
 mod kinds;
 mod reading;
 mod rules;
-mod rules_bypass;
 mod rules_fetch;
 mod rules_shapes;
 mod scoring;
