+++ crates/core/tests/quality/rules.rs
@@ -218,6 +218,297 @@ fn naming_a_projects_own_env_file_is_not_a_finding() {
 }
 
 #[test]
+fn safety_bypass_separates_a_switch_that_disables_a_check_from_prose_about_one() {
+    let switch = document(ItemKind::Skill, "commit with git commit --no-verify\n");
+    assert_eq!(
+        severity_of(&switch, "safety-bypass"),
+        Some(Severity::Critical)
+    );
+
+    let prose = document(
+        ItemKind::Skill,
+        "You may bypass safety when the build is green\n",
+    );
+    assert_eq!(severity_of(&prose, "safety-bypass"), Some(Severity::High));
+}
+
+/// Flags that ordinary tools carry say nothing on their own. The kendex
+/// `github` skill uses `--force` forty-two times, every one of them about
+/// its own documented override, and `--yes` is in every non-interactive
+/// install line there is.
+#[test]
+fn safety_bypass_leaves_ordinary_tool_flags_alone() {
+    let result = document(
+        ItemKind::Skill,
+        "git push --force-with-lease\napt install --yes ripgrep\n  --force          Skip checks\n",
+    );
+    assert!(
+        !rules_hit(&result).contains(&"safety-bypass"),
+        "{:?}",
+        result.findings
+    );
+}
+
+/// A document naming the switch in a code span is naming it. The control
+/// is the second case, and it is what the rule is worth: a fenced block is
+/// how a skill writes the command it means, so the switch inside one
+/// counts at full weight.
+#[test]
+fn safety_bypass_reads_a_switch_named_in_a_code_span_as_a_mention() {
+    let named = skill(&[("SKILL.md", "The bypass is `git commit --no-verify`.\n")]);
+    assert!(
+        !rules_hit(&named).contains(&"safety-bypass"),
+        "{:?}",
+        named.findings
+    );
+
+    let run = skill(&[("SKILL.md", "```sh\ngit commit --no-verify -m done\n```\n")]);
+    assert_eq!(
+        severity_of(&run, "safety-bypass"),
+        Some(Severity::Critical),
+        "{:?}",
+        run.findings
+    );
+}
+
+/// One named mention must not cover a use standing beside it: the rule
+/// reads every occurrence on the line, not the first.
+#[test]
+fn safety_bypass_reads_past_a_mention_to_the_use_behind_it() {
+    let result = skill(&[(
+        "SKILL.md",
+        "Never `--no-verify`, though this line says --no-verify plainly.\n",
+    )]);
+    assert_eq!(
+        severity_of(&result, "safety-bypass"),
+        Some(Severity::Critical),
+        "{:?}",
+        result.findings
+    );
+}
+
+/// A backtick that closes nothing quotes nothing. Markdown ends a run of
+/// backticks only on a run of the same length, so an opener with no match
+/// is literal text — and treating it as a toggle would let one stray
+/// backtick hide every switch after it on the line, which is this rule
+/// going quiet in a score somebody installs on.
+#[test]
+fn safety_bypass_still_reads_a_switch_after_a_backtick_that_closes_nothing() {
+    let result = skill(&[(
+        "SKILL.md",
+        "Prefer `git commit -m done over git commit --no-verify -m done.\n",
+    )]);
+    assert_eq!(
+        severity_of(&result, "safety-bypass"),
+        Some(Severity::Critical),
+        "{:?}",
+        result.findings
+    );
+}
+
+/// A span opened by two backticks holds single ones as its own text, which
+/// is how markdown writes a backtick at all. Reading the inner one as the
+/// close reports the rest of the span as though it stood in the open.
+#[test]
+fn safety_bypass_leaves_a_backtick_quoted_inside_a_longer_span() {
+    let result = skill(&[(
+        "SKILL.md",
+        "The shape is ``git commit` --no-verify`` in a doc.\n",
+    )]);
+    assert!(
+        !rules_hit(&result).contains(&"safety-bypass"),
+        "{:?}",
+        result.findings
+    );
+}
+
+/// The line an indented block opens on is the block's own text, the same
+/// as the line under it. Reading it as prose let a backtick there quote
+/// what the block prints literally, so a switch on a block's first line
+/// scored nothing and the same switch one line down scored Critical.
+///
+/// The control is the second case: in prose those same marks are
+/// markdown's, and they do quote.
+#[test]
+fn safety_bypass_reads_the_line_an_indented_block_opens_on_as_code() {
+    let opens = skill(&[("SKILL.md", "Run it:\n\n    git commit `--no-verify`\n")]);
+    assert_eq!(
+        severity_of(&opens, "safety-bypass"),
+        Some(Severity::Critical),
+        "{:?}",
+        opens.findings
+    );
+
+    let prose = skill(&[("SKILL.md", "Run it: git commit `--no-verify`\n")]);
+    assert!(
+        !rules_hit(&prose).contains(&"safety-bypass"),
+        "{:?}",
+        prose.findings
+    );
+}
+
+/// A code span closes on a later line of the same paragraph. Read one line
+/// at a time the opener meets no match on its own line and the close meets
+/// none on the next, so the switch quoted between them was reported as one
+/// standing in the open.
+///
+/// The control is the second case: the reach is the paragraph, not the
+/// document. A backtick whose partner stands past a blank line quotes
+/// nothing, and the switch beside it still counts.
+#[test]
+fn safety_bypass_leaves_a_switch_quoted_by_a_span_that_crosses_a_newline() {
+    let across = skill(&[(
+        "SKILL.md",
+        "The bypass is `git commit\n--no-verify`, which this never runs.\n",
+    )]);
+    assert!(
+        !rules_hit(&across).contains(&"safety-bypass"),
+        "{:?}",
+        across.findings
+    );
+
+    let apart = skill(&[(
+        "SKILL.md",
+        "The bypass is `git commit\n\n--no-verify` runs it.\n",
+    )]);
+    assert_eq!(
+        severity_of(&apart, "safety-bypass"),
+        Some(Severity::Critical),
+        "{:?}",
+        apart.findings
+    );
+}
+
+/// A heading is a block of its own, so a run of backticks left open at the
+/// end of one does not reach the paragraph under it. Joining the two into
+/// a single run paired those stray backticks, and the switch standing
+/// between them read as quoted: one unmatched backtick on each of two
+/// lines, and a line telling a reader to pass git the switch scored
+/// nothing.
+#[test]
+fn safety_bypass_reads_a_switch_a_heading_keeps_out_of_a_span() {
+    let result = skill(&[(
+        "SKILL.md",
+        "# Committing past the `hook\nPass --no-verify` when it complains.\n",
+    )]);
+    assert_eq!(
+        severity_of(&result, "safety-bypass"),
+        Some(Severity::Critical),
+        "{:?}",
+        result.findings
+    );
+}
+
+/// A heading underlined with `=` is the same block boundary written the
+/// other way round, and the underline closes it. A run reaching past that
+/// paired the two stray backticks and swallowed the line under it.
+#[test]
+fn safety_bypass_reads_a_switch_a_setext_heading_keeps_out_of_a_span() {
+    let result = skill(&[(
+        "SKILL.md",
+        "Committing past the `hook\n=========================\nPass --no-verify` when it complains.\n",
+    )]);
+    assert_eq!(
+        severity_of(&result, "safety-bypass"),
+        Some(Severity::Critical),
+        "{:?}",
+        result.findings
+    );
+}
+
+/// Each list item is a block of its own, so a backtick left open in one
+/// does not reach the next. A line carrying no marker continues the item
+/// above it, which is the second case: a span still crosses the newline
+/// inside one item.
+#[test]
+fn safety_bypass_reads_a_switch_a_list_item_keeps_out_of_a_span() {
+    let items = skill(&[(
+        "SKILL.md",
+        "- Write `git commit\n- Pass --no-verify` to git.\n",
+    )]);
+    assert_eq!(
+        severity_of(&items, "safety-bypass"),
+        Some(Severity::Critical),
+        "{:?}",
+        items.findings
+    );
+
+    let one = skill(&[(
+        "SKILL.md",
+        "- The bypass is `git commit\n  --no-verify`, which this never runs.\n",
+    )]);
+    assert!(
+        !rules_hit(&one).contains(&"safety-bypass"),
+        "{:?}",
+        one.findings
+    );
+}
+
+/// A blockquote is a block whether or not a blank line precedes it, so a
+/// backtick in the prose above one does not reach into it. Its findings
+/// still weigh one severity less, which is what a quotation is worth.
+///
+/// The control is the second case: two quoted lines are one block, so a
+/// span opened on the first still closes on the second.
+#[test]
+fn safety_bypass_reads_a_switch_a_blockquote_keeps_out_of_a_span() {
+    let into = skill(&[(
+        "SKILL.md",
+        "Write `git commit\n> Pass --no-verify` to git.\n",
+    )]);
+    assert_eq!(
+        severity_of(&into, "safety-bypass"),
+        Some(Severity::High),
+        "{:?}",
+        into.findings
+    );
+
+    let within = skill(&[(
+        "SKILL.md",
+        "> The bypass is `git commit\n> --no-verify`, which this never runs.\n",
+    )]);
+    assert!(
+        !rules_hit(&within).contains(&"safety-bypass"),
+        "{:?}",
+        within.findings
+    );
+}
+
+/// A backslash-escaped backtick is the character, not a delimiter. Reading
+/// it as one let a pair of them quote the switch between, which is the
+/// same silence one character smaller.
+#[test]
+fn safety_bypass_reads_a_switch_between_escaped_backticks() {
+    let result = skill(&[(
+        "SKILL.md",
+        "Write \\`git commit --no-verify\\` and run it.\n",
+    )]);
+    assert_eq!(
+        severity_of(&result, "safety-bypass"),
+        Some(Severity::Critical),
+        "{:?}",
+        result.findings
+    );
+}
+
+/// Switching an item off parks its file under a `.disabled` suffix and
+/// changes nothing about what the file is. A disabled artifact that scored
+/// differently would mean disabling something invented findings in it.
+#[test]
+fn a_disabled_markdown_artifact_scores_as_its_enabled_twin() {
+    let body = "The bypass is `git commit --no-verify`, which this never runs.\n";
+    let on = skill(&[("SKILL.md", body)]);
+    let off = skill(&[("SKILL.md.disabled", body)]);
+    assert!(
+        !rules_hit(&off).contains(&"safety-bypass"),
+        "{:?}",
+        off.findings
+    );
+    assert_eq!(rules_hit(&off), rules_hit(&on));
+    assert_eq!(off.safety.score, on.safety.score);
+}
+
+#[test]
 fn dangerous_commands_weigh_more_in_a_hook_than_in_a_skill() {
     let hook = audit(AuditInput {
         kind: ItemKind::Hook,
