+++ docs/ARCHITECTURE.md
@@ -781,10 +781,11 @@ lives in one capability table read by core and UI.
   read lossily and the replacements reported. A matched token never
   appears in any message, log or record — only a fingerprint.
 - **The file a harness loads is scanned at full weight, fences included.**
-  A fenced `sh` block in a SKILL.md *is* the instruction; a switch counts
-  where the line hands it to a program, quote marks and all — never in a
-  comment, a `case` pattern or a code span, and everywhere in a file whose
-  language this cannot read. One severity less for a blockquote and for a skill's supporting files (`tests/`, `fixtures/`, `references/`); secrets never weigh less.
+  A fenced `sh` block in a SKILL.md *is* the instruction, and a switch
+  counts wherever it stands as code rather than in a markdown code span.
+  One severity less for content that is plainly quoting: a blockquote, and
+  a skill's supporting files (`tests/`, `fixtures/`, `references/`).
+  Secrets never weigh less anywhere.
 - **Rule severities are calibrated against real catalogs.** Deobfuscation
   reports only what has no typographic use — invisible and bidirectional
   characters, letters imitating other letters — while normalizing emoji
